docker/requirements.txt
apache-airflow-providers-google==10.21.0
google-cloud-storage
google-cloud-bigquery
kaggle==1.6.12